+++ tests/test_match_complex.py
@@ -1,49 +1,49 @@
 """Tests for complex examples of MatchBox."""
 from matchbox.matchbox import MatchBox
 
-from tests import IndexedObject
+from tests import Entity
 
 
-def test_mixed_single_characteristic_objects():
+def test_mixed_single_characteristic_entities():
     """
     Test more complex example.
 
-    +--------+-------+-------------------------+
-    | Object | Match | Characteristic's values |
-    +========+=======+=========================+
-    | Ob1    | False | 1                       |
-    +--------+-------+-------------------------+
-    | Ob2    | True  | 3                       |
-    +--------+-------+-------------------------+
-    | Ob3    | False | 5                       |
-    +--------+-------+-------------------------+
-    | Ob4    | False |                         |
-    +--------+-------+-------------------------+
-    | Ob5    | True  | 1                       |
-    +--------+-------+-------------------------+
+    +--------+-------+------------------------+
+    | Entity | Match | Characteristic's trait |
+    +========+=======+========================+
+    | Ob1    | False | 1                      |
+    +--------+-------+------------------------+
+    | Ob2    | True  | 3                      |
+    +--------+-------+------------------------+
+    | Ob3    | False | 5                      |
+    +--------+-------+------------------------+
+    | Ob4    | False |                        |
+    +--------+-------+------------------------+
+    | Ob5    | True  | 1                      |
+    +--------+-------+------------------------+
 
     Will result in matchbox'es index:
 
-    +-----------+-----------------+
-    | Attribute | Matched Objects |
-    +===========+=================+
-    | 1         | Ob1, Ob2        |
-    +-----------+-----------------+
-    | 3         | Ob5             |
-    +-----------+-----------------+
-    | 5         | Ob2, Ob3, Ob5   |
-    +-----------+-----------------+
-    | Any new   | Ob2, Ob5        |
-    +-----------+-----------------+
+    +---------+------------------+
+    | Trait   | Matched Entities |
+    +=========+==================+
+    | 1       | Ob1, Ob2         |
+    +---------+------------------+
+    | 3       | Ob5              |
+    +---------+------------------+
+    | 5       | Ob2, Ob3, Ob5    |
+    +---------+------------------+
+    | Any new | Ob2, Ob5         |
+    +---------+------------------+
     """
-    # create objects
-    ob1 = IndexedObject(1, False)
-    ob2 = IndexedObject(3, True)
-    ob3 = IndexedObject(5, False)
-    ob4 = IndexedObject(None, False)
-    ob5 = IndexedObject(1, True)
+    # create entiities
+    ob1 = Entity(1, False)
+    ob2 = Entity(3, True)
+    ob3 = Entity(5, False)
+    ob4 = Entity(None, False)
+    ob5 = Entity(1, True)
 
-    # create matchbox and add objects
+    # create matchbox and add entities
     matchbox = MatchBox('characteristic')
     matchbox.add(ob1)
     matchbox.add(ob2)
@@ -78,12 +78,12 @@ def test_mixed_single_characteristic_objects():
     assert ob5 in matchbox.mismatch_unknown
 
 
-def test_mixed_multi_characteristic_objects():
+def test_mixed_multi_characteristic_entities():
     """
     Test more complex example.
 
     +--------+-------+-------------------------+
-    | Object | Match | Characteristic's values |
+    | Entity | Match | Characteristic's traits |
     +========+=======+=========================+
     | Ob1    | False | 1, 2                    |
     +--------+-------+-------------------------+
@@ -98,34 +98,34 @@ def test_mixed_multi_characteristic_objects():
 
     Should result in this index:
 
-    +-----------+-----------------+
-    | Attribute | Matched Objects |
-    +===========+=================+
-    | 1         | Ob1, Ob2        |
-    +-----------+-----------------+
-    | 2         | Ob1, Ob2, Ob5   |
-    +-----------+-----------------+
-    | 3         | Ob5             |
-    +-----------+-----------------+
-    | 4         | Ob5             |
-    +-----------+-----------------+
-    | 5         | Ob2, Ob3, Ob5   |
-    +-----------+-----------------+
-    | 6         | Ob2, Ob3, Ob5   |
-    +-----------+-----------------+
-    | 7         |                 |
-    +-----------+-----------------+
-    | Any other | Ob2, Ob5        |
-    +-----------+-----------------+
+    +-----------+------------------+
+    | Trait     | Matched Entities |
+    +===========+==================+
+    | 1         | Ob1, Ob2         |
+    +-----------+------------------+
+    | 2         | Ob1, Ob2, Ob5    |
+    +-----------+------------------+
+    | 3         | Ob5              |
+    +-----------+------------------+
+    | 4         | Ob5              |
+    +-----------+------------------+
+    | 5         | Ob2, Ob3, Ob5    |
+    +-----------+------------------+
+    | 6         | Ob2, Ob3, Ob5    |
+    +-----------+------------------+
+    | 7         |                  |
+    +-----------+------------------+
+    | Any other | Ob2, Ob5         |
+    +-----------+------------------+
     """
-    # create objects
-    ob1 = IndexedObject([1, 2], False)
-    ob2 = IndexedObject([3, 4, 7], True)
-    ob3 = IndexedObject([5, 6], False)
-    ob4 = IndexedObject(None, False)
-    ob5 = IndexedObject([1, 7], True)
+    # create entities
+    ob1 = Entity([1, 2], False)
+    ob2 = Entity([3, 4, 7], True)
+    ob3 = Entity([5, 6], False)
+    ob4 = Entity(None, False)
+    ob5 = Entity([1, 7], True)
 
-    # create matchbox and add objects
+    # create matchbox and add entities
     matchbox = MatchBox('characteristic')
     matchbox.add(ob1)
     matchbox.add(ob2)